test(usage): one test file per module, as the coverage gate requires
The gate's oracle is a strict sibling path (src/X.ts -> src/X.test.ts), so a
single src/usage.test.ts covering three modules counted as three untested files
and failed CI while passing locally. Split into src/usage-report.test.ts,
src/usage-prove.test.ts and src/commands/usage.test.ts, each documenting the
property it pins: unknown is never zero, a refusal only means something if the
harmless case is still allowed, and the box geometry must add up.
